requirements.txt
requests
python-dotenv
ipython
any-parser
markdown
beautifulsoup4
pymupdf
pillow
